add offset property to classes

+++ src/main/java/com/afunproject/dawncraft/classes/integration/epicfight/client/EpicFightPlayerRenderer.java
@@ -1,5 +1,6 @@
 package com.afunproject.dawncraft.classes.integration.epicfight.client;
 
+import com.afunproject.dawncraft.classes.data.DCClass;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Quaternion;
@@ -28,7 +29,7 @@ public class EpicFightPlayerRenderer {
         playerpatch = new AnimationPlayerPatch(player);
     }
     
-    public void render(PoseStack posestack, int x, int y, float partialTicks, String animation) {
+    public void render(PoseStack posestack, float x, float y, float partialTicks, String animation) {
         ClientAnimator animator = playerpatch.getClientAnimator();
         StaticAnimation anim = EpicFightMod.getInstance().animationManager.findAnimationByPath(animation);
         if (anim == null) anim = Animations.BIPED_IDLE;
